+++ flask_postgres.py
@@ -1,0 +1,18 @@
+import psycopg2
+from flask import Flask#,render_template
+
+
+app = Flask(__name__)
+con = psycopg2.connect(database="Movie", user="postgres", password="postgres", host="127.0.0.1", port="5000")##
+cursor = con.cursor()
+
+@app.route('/', methods =['post', 'get'])
+def testing():
+    cursor.execute("select * from actor")
+    result = cursor.fetchall
+    print(result)
+    
+    ##return render_template("testing.csv", data=result)
+
+if __name__ == '__main__':
+    app.run(debug=True)
